pyproject.toml
[tool.poetry]
name = "dynamictool"
version = "0.1.0"
description = ""
authors = ["mariyam <aic.ds.2022@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.11"
uvicorn = "^0.34.0"
psycopg = "^3.2.6"
psycopg2-binary = "^2.9.10"
sqlalchemy = "^2.0.39"
python-dotenv = "^1.0.1"
passlib = "^1.7.4"
python-jose = "^3.4.0"
pydantic = "^2.10.6"
email-validator = "^2.2.0"
python-multipart = "^0.0.20"
pandas = "^2.2.3"
aiosqlite = "^0.21.0"
bcrypt = "^4.3.0"
asyncpg = "^0.30.0"
pint = "^0.24.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
